requirements.txt
langchain>=0.2.0
langchain-google-genai>=0.2.0
langgraph>=0.2.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0
streamlit>=1.28.0 